requirements.txt
# Core logic
pandas>=2.0.0
numpy>=1.23

# PDF processing
pdfplumber>=0.10.0
PyMuPDF>=1.22.0
pdf2image>=1.16.3

# OCR
pytesseract>=0.3.10
Pillow>=9.5.0

# LLM (Google Gemini)
google-generativeai>=0.3.2
python-dotenv>=1.0.0

# Google Sheets integration
gspread>=5.7.2
oauth2client>=4.1.3

# Excel writing
openpyxl>=3.1.2